+++ src/frontend/packages/core/src/shared/components/list/list-types/endpoint/endpoint-card/endpoint-card.component.ts
@@ -1,13 +1,15 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { ReplaySubject, of as observableOf } from 'rxjs';
+import { ReplaySubject } from 'rxjs';
+
 import { EndpointModel } from '../../../../../../../../store/src/types/endpoint.types';
-import { EndpointsService } from '../../../../../../core/endpoints.service';
-import { CardStatus } from '../../../../../shared.types';
-import { CardCell } from '../../../list.types';
 import { UserFavoriteEndpoint } from '../../../../../../../../store/src/types/user-favorites.types';
+import { EndpointsService } from '../../../../../../core/endpoints.service';
 import { getFavoriteFromEndpointEntity } from '../../../../../../core/user-favorite-helpers';
-import { getNameForEndpointType, getFullEndpointApiUrl } from '../../../../../../features/endpoints/endpoint-helpers';
+import { getFullEndpointApiUrl, getNameForEndpointType } from '../../../../../../features/endpoints/endpoint-helpers';
+import { CardStatus } from '../../../../../shared.types';
 import { MetaCardMenuItem } from '../../../list-cards/meta-card/meta-card-base/meta-card.component';
+import { CardCell } from '../../../list.types';
+import { EndpointListHelper } from '../endpoint-list.helpers';
 
 @Component({
   selector: 'app-endpoint-card',
@@ -17,8 +19,9 @@ import { MetaCardMenuItem } from '../../../list-cards/meta-card/meta-card-base/m
 export class EndpointCardComponent extends CardCell<EndpointModel> implements OnInit {
 
   public status$ = new ReplaySubject<CardStatus>();
+  public rowObs = new ReplaySubject<EndpointModel>();
   public favorite: UserFavoriteEndpoint;
-  public type:  string;
+  public type: string;
   public address: string;
   public cardMenu: MetaCardMenuItem[];
 
@@ -29,35 +32,26 @@ export class EndpointCardComponent extends CardCell<EndpointModel> implements On
     this.type = row ? getNameForEndpointType(row.cnsi_type) : '';
     this.address = row ? getFullEndpointApiUrl(row) : '';
     this.status$.next(this.determineEndpointStatus(row));
+    this.rowObs.next(row);
   }
   get row(): EndpointModel {
     return this.pRow;
   }
 
-  constructor() {
+  constructor(
+    endpointListHelper: EndpointListHelper
+  ) {
     super();
-    this.cardMenu = [
-      {
-        label: 'Delete',
-        action: null,
-        can: observableOf(true)
-      },
-      {
-        label: 'Connect',
-        action: null,
-        can: observableOf(true)
-      },
-      {
-        label: 'Disconnect',
-        action: null,
-        can: observableOf(true)
-      }
-    ];
-   }
+    this.cardMenu = endpointListHelper.endpointActions().map(endpointAction => ({
+      label: endpointAction.label,
+      action: () => endpointAction.action(this.pRow),
+      can: endpointAction.createVisible(this.rowObs)
+    }));
+  }
 
-   ngOnInit() {
+  ngOnInit() {
     this.favorite = getFavoriteFromEndpointEntity(this.row);
-   }
+  }
 
   getLinkForEndpoint() {
     return EndpointsService.getLinkForEndpoint(this.row);
@@ -66,13 +60,13 @@ export class EndpointCardComponent extends CardCell<EndpointModel> implements On
   determineEndpointStatus(row: EndpointModel): CardStatus {
     switch (row.connectionStatus) {
       case 'connected':
-      return CardStatus.OK;
+        return CardStatus.OK;
       case 'disconnected':
-      return CardStatus.TENTATIVE;
+        return CardStatus.TENTATIVE;
       case 'checking':
-      return CardStatus.BUSY;
+        return CardStatus.BUSY;
       default:
-      return CardStatus.INCOMPLETE;
+        return CardStatus.INCOMPLETE;
     }
   }
 }
